Fix the error message for force_pkce
Bug
When the `force_pkce` option is enabled for a public client, and the
authorization code request doesn't include a code_challenge, the
error message returned is

"The authorization server encountered an unexpected condition which prevented it from fulfilling the request."

Root cause
PR https://github.com/doorkeeper-gem/doorkeeper/pull/1705 didn't add
the error message in the locales.

Fix
Base on RFC https://datatracker.ietf.org/doc/html/rfc7636#section-4.4.1,
we should return an "invalid_request" error with a proper description.
I added a new entry for the error message and included a test to verify
the proper message is returned.