Introspecting a token should behave the same wrt "revoke on use"
Our ResourceServer is a separate application from the AuthorizationServer,
so we can't modify our controllers to invoke `doorkeeper_authorize!`. Instead,
we make an API call to `token/introspect` on the provided token. The RS is
itself a client application; it uses `client_credentials` and has a special
`introspect` scope to allow it to examine tokens issued by any other client.
We noticed that revoke-on-use does not happen as it would if we could use
the provided `doorkeeper_authorize!` action helper.

We've been able to replicate the behavior in a test under the TokensController
spec.

We'd like to ask around about if this is behaving as designed, or something
configurable or a bug?